new onFailureToFetchSupergraphSdlDuringInit strategy
this change:
* removes old nginx proxy strategy
* simplifies down to the more complete aws-based strategy
* switches to using the new onFailureToFetchSupergraphSdlDuringInit method on UplinkSupergraphManager that will be public in a future version of apollo/gateway
* upgrades all dependencies
