+++ gateway/stream.js
