+++ infra/bin/infra.js
@@ -4,7 +4,7 @@ import cdk from "aws-cdk-lib";
 import { InfraStack } from "../lib/infra-stack.js";
 
 const app = new cdk.App();
-new InfraStack(app, "InfraStack", {
+new InfraStack(app, "ApolloManagedFederationResiliency", {
   /* If you don't specify 'env', this stack will be environment-agnostic.
    * Account/Region-dependent features and context lookups will not work,
    * but a single synthesized template can be deployed anywhere. */
