+++ infra/package.json
@@ -12,14 +12,14 @@
     "test": "jest"
   },
   "devDependencies": {
-    "aws-cdk": "2.3.0",
-    "jest": "^27.4.7"
+    "aws-cdk": "2.27.0",
+    "jest": "^28.1.1"
   },
   "dependencies": {
-    "@aws-cdk/aws-apigatewayv2-alpha": "^2.3.0-alpha.0",
-    "@aws-cdk/aws-apigatewayv2-integrations-alpha": "^2.3.0-alpha.0",
-    "aws-cdk-lib": "^2.3.0",
-    "constructs": "^10.0.25",
-    "typescript": "^4.5.4"
+    "@aws-cdk/aws-apigatewayv2-alpha": "^2.27.0-alpha.0",
+    "@aws-cdk/aws-apigatewayv2-integrations-alpha": "^2.27.0-alpha.0",
+    "aws-cdk-lib": "^2.27.0",
+    "constructs": "^10.1.37",
+    "typescript": "^4.7.3"
   }
 }
