--- nginx/docker-compose.yaml
@@ -1,31 +1,0 @@
-services:
-  gateway:
-    build: ./gateway
-    volumes:
-      - ./gateway:/web
-    ports:
-      - "4000:4000"
-    working_dir: /web
-    environment:
-      - PORT=4000
-      - APOLLO_SCHEMA_CONFIG_DELIVERY_ENDPOINT=http://proxy
-    command: yarn dev
-    depends_on:
-      - proxy
-
-  proxy:
-    image: nginx:latest
-    volumes:
-      - ./nginx/proxy.conf:/etc/nginx/nginx.conf
-      - ./tmp/:/data/nginx/cache
-    ports:
-      - "8080:80"
-    depends_on:
-      - apibreaker
-
-  apibreaker:
-    image: nginx:latest
-    volumes:
-      - ./nginx/apibreaker.conf:/etc/nginx/nginx.conf
-    ports:
-      - "8090:80"
