--- nginx/gateway/.env.sample
@@ -1,3 +1,0 @@
-APOLLO_KEY=
-APOLLO_GRAPH_VARIANT=current
-NODE_ENV=development
