--- nginx/nginx/proxy.conf
@@ -1,40 +1,0 @@
-worker_processes 2;
-
-events {
-  worker_connections 1024;
-}
-
-http {
-  client_max_body_size 100M;
-  
-  proxy_cache_path /data/nginx/cache levels=1:2 keys_zone=my_cache:10m max_size=10g 
-                 inactive=60m use_temp_path=off;
-
-  server {
-    listen 80;
-    resolver 127.0.0.11;
-
-    location / {
-      proxy_cache my_cache;
-
-      # The composition-config-link endpoint responds with 
-      # `cache-control: private, max-age=0, no-transform` so nginx will not 
-      # cache it and it won't be available for `proxy_cache_use_stale`. By 
-      # ignoring the headers and caching all 200 responses for some period of 
-      # time, the response ends up on disk and `proxy_cache_use_stale` will use 
-      # it when the upstream is unavailable.
-      proxy_ignore_headers Cache-Control;
-      proxy_cache_valid 200 1s;
-
-      proxy_http_version 1.1;
-      proxy_cache_revalidate on;
-      proxy_cache_methods GET HEAD POST;
-      proxy_cache_use_stale error timeout updating http_500 http_502 http_503 http_504;
-      proxy_cache_background_update on;
-      proxy_cache_lock on;
-
-      # proxy_pass https://uplink.api.apollographql.com;
-      proxy_pass http://apibreaker:80;
-    }
-  }
-}
