--- webhook-aws/gateway/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "gateway",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@apollo/gateway": "^0.45.0",
-    "@aws-sdk/client-secrets-manager": "^3.45.0",
-    "apollo-server-lambda": "^3.6.1",
-    "esbuild": "^0.14.10",
-    "graphql": "^15.7.2",
-    "typescript": "^4.5.4"
-  },
-  "scripts": {
-    "build": "esbuild --bundle index.js --outdir=out --platform=node --target=node14 --minify"
-  }
-}
