--- webhook-aws/infra/cdk.context.json
@@ -1,4 +1,0 @@
-{
-  "webhookKeyArn": "",
-  "apolloKeyArn": ""
-}
