--- webhook-aws/webhook-handler/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "webhook-handler",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@aws-sdk/client-s3": "^3.45.0",
-    "@aws-sdk/client-secrets-manager": "^3.45.0",
-    "esbuild": "^0.14.10",
-    "make-fetch-happen": "^9.1.0",
-    "typescript": "^4.5.4"
-  },
-  "scripts": {
-    "build": "esbuild --bundle index.js --outdir=out --platform=node --target=node14 --minify"
-  },
-  "devDependencies": {
-    "@types/aws-lambda": "^8.10.89",
-    "@types/make-fetch-happen": "^9.0.1",
-    "@types/node": "^17.0.8"
-  }
-}
