+++ webhook-handler/index.js
@@ -1,15 +1,10 @@
-import {
-  S3Client,
-  PutObjectCommand,
-  GetObjectCommand,
-} from "@aws-sdk/client-s3";
+import { S3Client, PutObjectCommand } from "@aws-sdk/client-s3";
 import {
   SecretsManagerClient,
   GetSecretValueCommand,
 } from "@aws-sdk/client-secrets-manager";
 import { createHmac } from "crypto";
 import fetch from "make-fetch-happen";
-import { streamToString } from "./stream.js";
 
 const client = new S3Client({
   region: process.env.BUCKET_REGION,
@@ -75,84 +70,6 @@ export async function webhookHandler(event) {
   return { statusCode: 200, body: '{ "ok": true }' };
 }
 
-/**
- * Function handler for a Uplink fallback API. It retrieves the supergraph from
- * the S3 bucket and mimics the Uplink API response to act as a drop-in replacement.
- * @type {import("aws-lambda").Handler<import("aws-lambda").APIGatewayProxyEvent, import("aws-lambda").APIGatewayProxyResult>}
- */
-export async function fallbackHandler(event) {
-  const apiKey = await getApolloKey();
-
-  if (!event.body) {
-    return errorResponse("Missing request body");
-  }
-
-  /** @type {RequestShape} */
-  const { variables } = JSON.parse(event.body);
-
-  if (variables.apiKey !== apiKey) {
-    return errorResponse("Invalid API key");
-  }
-
-  const objectKey = `${variables.ref}.graphql`;
-
-  try {
-    const obj = await client.send(
-      new GetObjectCommand({
-        Bucket: process.env.BUCKET_NAME,
-        Key: objectKey,
-      })
-    );
-
-    if (!obj || !obj.Body) {
-      return errorResponse("Schema not found");
-    }
-
-    return {
-      statusCode: 200,
-      headers: {
-        "content-type": "application/json",
-      },
-      body: JSON.stringify({
-        data: {
-          routerConfig: {
-            __typename: "RouterConfigResult",
-            id: obj.Metadata?.["last-modified"] ?? "0",
-            supergraphSdl: await streamToString(obj.Body),
-          },
-        },
-      }),
-    };
-  } catch (/** @type {any} */ error) {
-    if (error.message === "NoSuchKey") {
-      return errorResponse(`${objectKey} not found`);
-    } else {
-      return errorResponse(error.message);
-    }
-  }
-}
-
-/**
- * @param {string} message
- */
-function errorResponse(message) {
-  return {
-    statusCode: 200,
-    headers: {
-      "content-type": "application/json",
-    },
-    body: JSON.stringify({
-      data: {
-        routerConfig: {
-          __typename: "FetchError",
-          code: "ERROR_FROM_FALLBACK",
-          message,
-        },
-      },
-    }),
-  };
-}
-
 // ****************************************************************************
 // secrets management for the webhook signing secret and Apollo API key
 // ****************************************************************************
@@ -173,22 +90,6 @@ async function getWebhookSecret() {
   }
 
   return memoizedWebhookSecretValue;
-}
-
-/** @type {string | undefined} */
-let memoizedApolloKeySecretValue;
-
-async function getApolloKey() {
-  if (!memoizedApolloKeySecretValue) {
-    const command = new GetSecretValueCommand({
-      SecretId: process.env.APOLLO_KEY_ARN,
-    });
-
-    const resp = await secretsClient.send(command);
-    memoizedApolloKeySecretValue = resp.SecretString;
-  }
-
-  return memoizedApolloKeySecretValue;
 }
 
 /**
