+++ webhook-handler/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "webhook-handler",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "@aws-sdk/client-s3": "^3.107.0",
+    "@aws-sdk/client-secrets-manager": "^3.105.0",
+    "esbuild": "^0.14.43",
+    "make-fetch-happen": "^10.1.7",
+    "typescript": "^4.7.3"
+  },
+  "scripts": {
+    "build": "esbuild --bundle index.js --outdir=out --platform=node --target=node14 --minify"
+  },
+  "devDependencies": {
+    "@types/aws-lambda": "^8.10.100",
+    "@types/make-fetch-happen": "^9.0.2",
+    "@types/node": "^17.0.42"
+  }
+}
