--- webhook/apibreaker.conf
@@ -1,19 +1,0 @@
-worker_processes 2;
-
-events {
-  worker_connections 1024;
-}
-
-http {
-  client_max_body_size 100M;
-
-  server {
-    listen 80;
-    resolver 127.0.0.11;
-
-    location / {
-      # return 500 "<b>Down!</b>";
-      proxy_pass https://uplink.api.apollographql.com;
-    }
-  }
-}
