--- webhook/cache-service/Dockerfile
@@ -1,6 +1,0 @@
-FROM node:14
-WORKDIR /web
-COPY package.json yarn.lock /web
-RUN yarn
-COPY . /web
-CMD yarn start
