--- webhook/cache-service/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "webhook",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "type": "module",
-  "dependencies": {
-    "@types/body-parser": "^1.19.0",
-    "@types/express": "^4.17.12",
-    "@types/make-fetch-happen": "^8.0.1",
-    "@types/morgan": "^1.9.2",
-    "@types/node": "^15.12.0",
-    "express": "^4.17.1",
-    "make-fetch-happen": "^9.0.2",
-    "morgan": "^1.10.0",
-    "nodemon": "^2.0.7",
-    "typescript": "^4.3.2"
-  },
-  "scripts": {
-    "dev": "nodemon index.js"
-  }
-}
