--- webhook/cache-service/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "es2020",
-    "checkJs": true,   
-    "noEmit": true,   
-    "strict": true,   
-    "moduleResolution": "node",
-    "esModuleInterop": true,   
-    "skipLibCheck": true,      
-    "forceConsistentCasingInFileNames": true
-  }
-}
