--- webhook/docker-compose.yaml
@@ -1,34 +1,0 @@
-services:
-  gateway:
-    build: ./gateway
-    volumes:
-      - ./gateway:/web
-    ports:
-      - "4000:4000"
-    working_dir: /web
-    environment:
-      - PORT=4000
-      - APOLLO_SCHEMA_CONFIG_DELIVERY_ENDPOINT=http://apibreaker
-      - SDL_BACKUP_ENDPOINT=http://cache-service:5000/backup-sdl
-    command: yarn dev
-    depends_on:
-      - cache-service
-      - apibreaker
-
-  cache-service:
-    build: ./cache-service
-    volumes:
-      - ./cache-service:/web
-    ports:
-      - "5000:5000"
-    working_dir: /web
-    environment:
-      - PORT=5000
-    command: yarn dev
-
-  apibreaker:
-    image: nginx:latest
-    volumes:
-      - ./apibreaker.conf:/etc/nginx/nginx.conf
-    ports:
-      - "8090:80"
