--- webhook/gateway/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "gateway",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "type": "module",
-  "dependencies": {
-    "@apollo/gateway": "^0.28.2",
-    "apollo-server": "^2.25.0",
-    "dotenv": "^10.0.0",
-    "graphql": "^15.5.0",
-    "make-fetch-happen": "^9.0.1",
-    "nodemon": "^2.0.7"
-  },
-  "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon -r dotenv/config index.js"
-  }
-}
